### 2020-10-07 #2728 @whedon
[PRE REVIEW]: OpenSCM Two Layer Model: A Python implementation of the two-layer climate model
**Submitting author:** @znicholls (<a href="http://orcid.org/0000-0002-4767-2723">Zebedee Nicholls</a>)
**Repository:** <a href="https://github.com/openscm/openscm-twolayermodel" target ="_blank">https://github.com/openscm/openscm-twolayermodel</a>
**Version:** v0.1.2
