### 2020-10-07 #2728 @znicholls
> In addition, [this list of people](https://bit.ly/joss-reviewers)

Looking through the list, these people list climate or atmospheric science as areas so could be willing

- DamienIrving
- jtmiclat
- ritwikagarwal
- koldunovn
- eviatarbach
- arbennett
- bradyrx
- OmarShehata
- marvinjonathcn
- rchg
- andreas-h
- dhhagan
- josephhardinee
- ashwinvis
- ali-ramadhan
- rchg
- capitalaslash
- slayoo
- simonom
- rabernat
- sadielbartholomew
- sarats
