### 2020-10-07 #2728 @znicholls
To the editors, I've just read the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort) properly. Apologies that I missed the description of effort! I am very happy to close this request if it appears to be out of scope because of its limited size. Is it possible to reopen if/when we've been working on the implementation for more time and added more features?
