### 2020-10-12 fe73599
Editorial fixes to the JOSS paper
Made a few style fixes (mostly using " instead of the LaTex style backtick quotes).

Related to openjournals/joss-reviews#2728
