### 2020-10-12 ff538c3
Fix broken ReadTheDocs links in the README
The link to the main docs and the contributing guide were broken because
of missing `-` in the project name part of the URL. 

Related to openjournals/joss-reviews#2728
