### 2020-10-12 #2728 @leouieda
👋🏽 Hi @znicholls thank you for the review suggestions!

>  I am very happy to close this request if it appears to be out of scope because of its limited size. Having said that, this implementation (albeit in a different repository at the time) has already been used in the academic literature and is likely to be used again in future given the prominence of this model.

I had a look at the software repository and the paper. This does seem more than ["3 months of work"](https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort) to me given that you have invested in a decent amount of testing and the model has been used in research already. Bare in mind that I'm not a climate scientist, though. We can investigate further if any reviewers have concerns but otherwise I'm happy to proceed with the review.

Please have a look at the "missing DOIs" report above and add the relevant DOIs to your `paper.bib`. I have also submitted a PR with some quick fixes to the paper format.
