### 2020-10-12 #2728 @leouieda
:wave: @sadielbartholomew @kdorheim @bradyrx would you be willing to review this submission for [JOSS](https://joss.theoj.org/)? Reviews are done entirely through GitHub (issues and pull requests). These are the reviewer guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

This is a pre-review issue in which I will assign reviewers. Once there are sufficient reviewers (usually 2-3), I will open a new issue where the review will take place. If you agree to review, you will be notified when the review process if formally started.

If you are not able to review at this time, please let me know if you have recommendations for other reviewers. **Please recommend without mentioning a handle to avoid notifying the person** (e.g., you would refer to me as leouieda instead of @leouieda). This way we attempt to reduce the number of people who are subscribed to this thread.
