### 2020-10-12 #2728 @kdorheim
leouieda unfortunately I am unable to review this at the moment given other time commitments. I am not sure if he is registered to review for JOSS but Alexey Shimnokol (https://github.com/ashiklom) comes to mind as someone who has worked on SCMs, is a good reviewer, a python user, and champion of open source science.
