### 2020-10-12 #2728 @sadielbartholomew
Thanks leouieda. I am very busy this October but in November I will have a lot more free time. I am happy to review assuming the editor and authors wouldn't mind waiting on my review until early November, but that is still a number of weeks away so I appreciate that may be longer than reasonable or desired to wait.

So feel free to assign me to review or not depending on timescales required. Certainly if someone else is free to review more quickly, it would be better to choose them.
