### 2020-10-12 #2728 @leouieda
:wave: @ashiklom would you be willing to review this submission for [The Journal of Open Source Software (JOSS)](https://joss.theoj.org/)? JOSS is a developer-friendly open-access journal and reviews are done entirely through GitHub (issues and pull requests). These are our reviewer guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html Please see the previous comments in this thread for more context. The software submission is https://github.com/openscm/openscm-twolayermodel Thank you for your time 🙂
