### 2020-10-22 #2728 @leouieda
@ashiklom @sadielbartholomew if this is your first time reviewing for JOSS, you should get an invitation to join the `openjournals` GitHub organization (usually by email). Please accept the invitation so that you can edit your checklists in the JOSS review. The invitation will expire so it's best to do this as soon as possible. If you have any issues, we can issue the invitation again (just let me know).
