### 2020-10-23 openscm/openscm-twolayermodel#16 @ashiklom
Quote notebook install instructions in README
`bash` doesn't seem to mind hard brackets, but `zsh` (note: the default shell on recent MacOS) doesn't like them in unquoted form and throws the following error:

```
zsh: no matches found: openscm-twolayermodel[notebooks]
```

This is easily resolved by quoting the full package name:

```
pip install "openscm-twolayermodel[notebooks]"
```

I would suggest tweaking this in the installation docs.

(Discovered as part of https://github.com/openjournals/joss-reviews/issues/2766).
