### 2020-10-23 openscm/openscm-twolayermodel#17 @ashiklom
General Usage documentation
On the [General Usage](https://openscm-two-layer-model.readthedocs.io/en/latest/usage.html) page of the documentation, it would be useful to add a hyperlink to the notebooks directory of the repository (since GitHub does a nice job of rendering the notebooks).

Also / Alternatively, you might consider rendering the notebooks within the ReadTheDocs documentation -- it seems like it should be possible to do this automatically with [`nbsphinx`](https://nbsphinx.readthedocs.io/en/0.8.0/usage.html#Automatic-Creation-of-HTML-and-PDF-output-on-readthedocs.org), but I've never actually tried it.

(Opened as part of https://github.com/openjournals/joss-reviews/issues/2766).
