### 2020-10-26 openscm/openscm-twolayermodel#18 @ashiklom
Add statement of need to documentation home page
Per the JOSS recommendations, the documentation should address, "Do the authors clearly state what problems the software is designed to solve and who the target audience is?". I would recommend doing this in 1-3 sentences on the [documentation home page](https://openscm-two-layer-model.readthedocs.io/en/latest/index.html). The JOSS paper does this pretty well, so you could just lift the text directly from that.

(Part of JOSS review, https://github.com/openjournals/joss-reviews/issues/2766).
