### 2020-10-26 a7decd3
option to remove per-scenario progress bar
