### 2020-10-27 876f6ab
Apply suggestions from code review
Co-authored-by: Zeb Nicholls <zebedee.nicholls@climate-energy-college.org>
