### 2020-10-28 #2766 @ashiklom
## General checks

- **Repository**: Link is valid; code is well-organized according to typical standards of a Python module.
- **License**: BSD 3-caluse, present in full in [LICENSE file](https://github.com/openscm/openscm-twolayermodel/blob/master/LICENSE) in the repository
- **Contribution and authorship**: Submission author @znicholls is the [largest contributor](https://github.com/openscm/openscm-twolayermodel/graphs/contributors) by far. No obvious issues with the author list.
- **Substantial scholarly effort**:
  - Oldest commit is [April 15, 2020](https://github.com/openscm/openscm-twolayermodel/commit/85f9d050ab724b8b4c51d7412635bcaa1a726344).
  - This model is a new addition to the growing list of reduced-complexity climate models. It serves as an interesting alternative to similar efforts, such as [Hector](https://github.com/jgcri/hector), [MAGICC](www.magicc.org), and [FAIR](https://github.com/OMS-NetZero/FAIR), and in this context, is a member of the Reduced Complexity Model Intercomparison Project (RCMIP), phase 1.

## Functionality

- **Installation**: I successfully installed the module and its dependencies via `pip` from the Python Package Index (PyPI) and from source, as described in the instructions.
- **Functionality**: I ran all of the notebooks without errors, and ran a few simple experiments of my own, tweaking parameters of the model to explore the response. Everything worked as expected.
- **Performance**: There are no performance claims. However, the notebooks run almost instantaneously on my 2019 Macbook Pro -- this is a fast model!

## Documentation

- **Statement of need**: Missing from documentation. I suggest adding a brief (1-3 sentences) note about this to the [documentation home page](https://openscm-two-layer-model.readthedocs.io/en/latest/index.html). See https://github.com/openscm/openscm-twolayermodel/issues/18.
- **Installation instructions**: Described in the README, and are very straightforward (see above).
- **Example usage**: Four Jupyter notebooks distributed with the repository provide useful examples of general usage and a few interesting specific cases.
- **Functionality documentation**: All methods and their arguments are well documented using ReadTheDocs.
- **Automated tests**: The installation process and Jupyter notebooks (examples of typical usage) are [automatically tested using GitHub Actions](https://github.com/openscm/openscm-twolayermodel/actions/runs/303096360/workflow).
- **Community guidelines**: The authors provide a thorough and detailed contributing guide in the ["Development" section of their documentation](https://openscm-two-layer-model.readthedocs.io/en/latest/development.html).

## Software paper

- **Summary**: Summary provides a brief overview of the role of simple climate models, a concise theoretical description of the two-layer model, and a description of the implementation here.
- **Statement of need**: Present, and provides a compelling case for the extensible nature of this implementation. Also makes a useful comparison with another similar model (FaIR). One minor suggestion: It may be worth advertising the fact that this implementation is in Python, a language that is open source, easy to use, and popular in the atmospheric sciences (and Earth and computational sciences more generally). For instance, were previous implementation of the two-layer model in something less user-friendly like Fortran or less open like Matlab?
- **State of the field**: Generally present in the Summary, but could be fleshed out a bit. Specifically, it might be worth (briefly) giving a few other simple climate model examples (that can be pulled directly from the RCMIP paper). Even more useful might be a short table stating the model, a brief description of its scientific basis (in a few words; <1 full sentence), and its programming language and paradigm. This would also provide useful context for this model.
- **Quality of writing**: Writing is solid; no major concerns.
- **References**: References are used appropriately and formatted correctly.
