### 2020-10-29 #2766 @leouieda
@ashiklom thank you for your thorough review (and issues on the software repository)!

@znicholls please do you best to address some of the points raised, in particular:

1. Adding suggested context to the statement of need. Including it in the docs is a good idea and helps define the scope of the project.
2. Fleshing out the "State of the field" in the JOSS paper.

Please keep us posted on your progress 🙂
