### 2020-10-31 4b6674a
Quote installation instructions in README, thanks @ashiklom
