### 2020-10-31 openscm/openscm-twolayermodel#16 @znicholls
Awesome thanks! I've put a fix in #21 if you want to take a look
