### 2020-10-31 fcd572a
Add statement of need to README
