### 2020-10-31 openscm/openscm-twolayermodel#18 @znicholls
Thanks again! Fix in #22
