### 2020-10-31 b667621
Ref JOSS review in CHANGELOG
