### 2020-10-31 #2766 @znicholls
> @znicholls please do you best to address some of the points raised, in particular

Will do. I've started in a couple of PRs, will work on 2 this week. Thanks to both for the great review and co-ordination!
