### 2020-10-31 a58fbae
Add JOSS badge to README
