### 2020-10-31 43dd298
Add first notebook to docs/source/usage
