+++ setup.py
@@ -32,7 +32,8 @@ REQUIREMENTS_TESTS = [
     "pytest>=4.0",
     "scipy",
 ]
-REQUIREMENTS_DOCS = ["sphinx>=1.4", "sphinx_rtd_theme", "sphinx-click"]
+REQUIREMENTS_DOCS = [
+    "nbsphinx", "sphinx>=1.4", "sphinx_rtd_theme", "sphinx-click","sphinx-copybutton"]
 REQUIREMENTS_DEPLOY = ["twine>=1.11.0", "setuptools>=38.6.0", "wheel>=0.31.0"]
 
 REQUIREMENTS_DEV = [
