+++ docs/source/conf.py
@@ -40,6 +40,9 @@ extensions = [
     "sphinx.ext.coverage",
     "sphinx.ext.intersphinx",
     "sphinx.ext.napoleon",  # pass numpy style docstrings
+    "nbsphinx",
+    "sphinx.ext.mathjax",
+    "sphinx_copybutton",
 ]
 autodoc_default_options = {
     "inherited-members": None,
@@ -71,7 +74,7 @@ language = "en"
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ["build"]
+exclude_patterns = ['build', '**.ipynb_checkpoints']
 
 # The name of the Pygments (syntax highlighting) style to use.
 pygments_style = "sphinx"
@@ -199,7 +202,13 @@ epub_exclude_files = ["search.html"]
 
 # -- Extension configuration -------------------------------------------------
 
+## cov
+
 coverage_write_headline = False  # do not write headlines.
+
+## nbsphinx
+
+nbsphinx_execute = 'auto'
 
 # -- Options for intersphinx extension ---------------------------------------
 
