+++ docs/source/usage.rst
@@ -1,7 +1,20 @@
 Usage
 =====
 
-.. contents:: Contents
-    :local:
+Here we provide examples of OpenSCM two layer model's behaviour and usage.
+The source code of these usage examples is available in the folder
+`docs/source/usage`_ of the `GitHub repository`_.
 
-All of our usage examples are included in ``openscm-twolayermodel/notebooks``.
+.. _`docs/source/usage`:
+   https://github.com/openscm/openscm-twolayermodel/tree/master/docs/source/usage
+
+.. _`GitHub repository`:
+   https://github.com/openscm/openscm-twolayermodel
+
+Basic demos
++++++++++++
+
+.. toctree::
+   :maxdepth: 1
+
+   usage/getting-started.ipynb
