+++ .github/workflows/ci-cd-workflow.yml
@@ -94,11 +94,11 @@ jobs:
         pip install -e .[tests,notebooks] --use-feature=2020-resolver
     - name: Test notebooks with nbval
       run: |
-          pytest notebooks -r a --nbval-lax --sanitize-with tests/notebook-tests.cfg --no-cov
+          pytest docs/source/usage -r a --nbval-lax --sanitize-with tests/notebook-tests.cfg --no-cov
     - name: Test notebooks strictly
       if: matrix.python-version == 3.7
       run: |
-        pytest notebooks -r a --nbval --sanitize-with tests/notebook-tests.cfg --no-cov
+        pytest docs/source/usage -r a --nbval --sanitize-with tests/notebook-tests.cfg --no-cov
 
   deploy-pypi:
     needs: test-notebooks
