### 2020-10-31 openscm/openscm-twolayermodel#17 @znicholls
Great suggestion thanks. I've done it, it means the notebooks move to a slightly less obvious place but the documentation is more complete so I think a good tradeoff?
