+++ docs/source/conf.py
@@ -74,7 +74,7 @@ language = "en"
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ['build', '**.ipynb_checkpoints']
+exclude_patterns = ["build", "**.ipynb_checkpoints"]
 
 # The name of the Pygments (syntax highlighting) style to use.
 pygments_style = "sphinx"
@@ -208,7 +208,7 @@ coverage_write_headline = False  # do not write headlines.
 
 ## nbsphinx
 
-nbsphinx_execute = 'always'
+nbsphinx_execute = "always"
 
 # -- Options for intersphinx extension ---------------------------------------
 
