### 2020-10-31 3f8d040
Add notebook dependency to docs environment
