### 2020-10-31 fe9a31c
Add ipywidgets extension
