### 2020-10-31 b916f77
Make docs also depend on notebook dependencies
