+++ .github/workflows/ci-cd-workflow.yml
@@ -28,8 +28,11 @@ jobs:
         pydocstyle src
         bandit -c .bandit.yml -r openscm_twolayermodel
         flake8 src tests setup.py
+    - name: Install pandoc
+      uses: r-lib/actions/setup-pandoc@v1
+      with:
+        pandoc-version: '2.11'
     - name: Build docs
-      uses: docker://pandoc/core:2.9
       # treat warnings as errors (-W)...
       # ...but not when being nitpicky (-n)
       run: |
