+++ paper.bib
@@ -87,7 +87,8 @@ year = {2015}
   number={1-2},
   pages={213},
   year={2011},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/s10584-011-0156-z}
 }
 
 @article{rohrschneider_2019_simple,
@@ -98,7 +99,8 @@ year = {2015}
   number={5-6},
   pages={3131--3145},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/s00382-019-04686-4}
 }
 
 @Article{smith_2018_fairv1_3,
