### 2020-11-02 #2766 @znicholls
I've updated the missing DOIs (https://github.com/openscm/openscm-twolayermodel/pull/24) so will recompile
