### 2020-11-02 77dd135
Add note about Python
