+++ paper.bib
@@ -134,3 +134,12 @@ DOI = {10.5194/gmd-11-2273-2018}
   URL = {https://gmd.copernicus.org/preprints/gmd-2019-375/},
   DOI = {10.5194/gmd-2019-375}
 }
+
+@misc{pint,
+  author = {H. E. Grecco and Pint Authors},
+  title = {Pint: Operate and manipulate physical quantities in Python},
+  year = {2020},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/hgrecco/pint}
+}
