### 2020-11-02 #2766 @znicholls
Text updates based on @ashiklom's review will go in https://github.com/openscm/openscm-twolayermodel/pull/25
