### 2020-11-02 a27a79e
Add table with examples of openly accessible models
