+++ paper.bib
@@ -125,21 +125,51 @@ DOI = {10.5194/gmd-11-2273-2018}
 }
 
 @Article{rcmip_phase_1,
-  AUTHOR = {Nicholls, Z. R. J. and Meinshausen, M. and Lewis, J. and Gieseke, R. and Dommenget, D. and Dorheim, K. and Fan, C.-S. and Fuglestvedt, J. S. and Gasser, T. and Gol\"uke, U. and Goodwin, P. and Kriegler, E. and Leach, N. J. and Marchegiani, D. and Quilcaille, Y. and Samset, B. H. and Sandstad, M. and Shiklomanov, A. N. and Skeie, R. B. and Smith, C. J. and Tanaka, K. and Tsutsui, J. and Xie, Z.},
-  TITLE = {Reduced complexity model intercomparison project phase 1: Protocol, results and initial observations},
-  JOURNAL = {Geoscientific Model Development Discussions},
-  VOLUME = {2020},
+  AUTHOR = {Nicholls, Z. R. J. and Meinshausen, M. and Lewis, J. and Gieseke, R. and Dommenget, D. and Dorheim, K. and Fan, C.-S. and Fuglestvedt, J. S. and Gasser, T. and Gol\"uke, U. and Goodwin, P. and Hartin, C. and Hope, A. P. and Kriegler, E. and Leach, N. J. and Marchegiani, D. and McBride, L. A. and Quilcaille, Y. and Rogelj, J. and Salawitch, R. J. and Samset, B. H. and Sandstad, M. and Shiklomanov, A. N. and Skeie, R. B. and Smith, C. J. and Smith, S. and Tanaka, K. and Tsutsui, J. and Xie, Z.},
+  TITLE = {Reduced Complexity Model Intercomparison Project Phase 1: introduction and evaluation of global-mean temperature response},
+  JOURNAL = {Geoscientific Model Development},
+  VOLUME = {13},
   YEAR = {2020},
-  PAGES = {1--33},
-  URL = {https://gmd.copernicus.org/preprints/gmd-2019-375/},
-  DOI = {10.5194/gmd-2019-375}
+  NUMBER = {11},
+  PAGES = {5175--5190},
+  URL = {https://gmd.copernicus.org/articles/13/5175/2020/},
+  DOI = {10.5194/gmd-13-5175-2020}
 }
 
 @misc{pint,
-  author = {H. E. Grecco and Pint Authors},
+  author = {H. E. Grecco and others},
   title = {Pint: Operate and manipulate physical quantities in Python},
   year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/hgrecco/pint}
+}
+
+@Article{Dommenget_2011_greb,
+  author={Dommenget, Dietmar
+  and Fl{\"o}ter, Janine},
+  title={Conceptual understanding of climate change with a globally resolved energy balance model},
+  journal={Climate Dynamics},
+  year={2011},
+  month={Dec},
+  day={01},
+  volume={37},
+  number={11},
+  pages={2143-2165},
[…]
