### 2020-11-02 #2766 @znicholls
@ashiklom I've added a table with openly accessible models from RCMIP (so just FaIR, GREB and Hector) to openscm/openscm-twolayermodel#25. Does that work or would extra text and explanation be useful?
