+++ CHANGELOG.rst
@@ -25,6 +25,7 @@ Added
 Changed
 ~~~~~~~
 
+- (`#25 <https://github.com/openscm/openscm-twolayermodel/pull/25>`_) JOSS paper following `JOSS review 1 <https://github.com/openjournals/joss-reviews/issues/2766#issuecomment-718025503>`_
 - (`#23 <https://github.com/openscm/openscm-twolayermodel/pull/23>`_) Moved notebooks into full documentation following `JOSS review <https://github.com/openjournals/joss-reviews/issues/2766>`_ (closes `#17 <https://github.com/openscm/openscm-twolayermodel/issues/17>`_)
 - (`#21 <https://github.com/openscm/openscm-twolayermodel/pull/21>`_) Quoted pip install instructions to ensure cross-shell compatibility following `JOSS review <https://github.com/openjournals/joss-reviews/issues/2766>`_ (closes `#16 <https://github.com/openscm/openscm-twolayermodel/issues/16>`_)
 - (`#20 <https://github.com/openscm/openscm-twolayermodel/pull/20>`_) Option to remove tqdm progress bar by passing ``progress=False``
