### 2020-11-02 #2766 @rgieseke
>>        For instance, were previous implementation of the two-layer model in something less user-friendly like Fortran or less open like Matlab?

> On this one I'm not sure. There's FaIR, but it is missing some of the conversion functionality and modularisation we've implemented here (as mentioned in the draft). I haven't been able to find any other public implementations so couldn't comment on other languages unfortunately.

Are there any openly available implementations? Maybe highlight that more, if that's the case?
