### 2020-11-02 #2766 @znicholls
> Are there any openly available implementations?

Not as far as I'm aware.
