### 2020-11-02 #2766 @rgieseke
> @ashiklom I've added a table with openly accessible models from RCMIP (so just FaIR, GREB and Hector) to [openscm/openscm-twolayermodel#25](https://github.com/openscm/openscm-twolayermodel/pull/25). Does that work or would extra text and explanation be useful?

Maybe one could reference models that are easily driven in forcing mode? Oscar, WASP and Pymagicc/MAGICC are also openly available on GitHub, but then one slowly ends up referencing the entire table from the RCMIP paper ...
