### 2020-11-02 openscm/openscm-twolayermodel#17 @ashiklom
Rendered notebooks look great! Seems like a fine trade-off to me, as long as the links to the notebooks are up to date everywhere.
