### 2020-11-02 #2766 @ashiklom
I think the table you added combined with the references to the RCMIP paper throughout the documentation are sufficient to address this (you might consider making it _extra_ clear that the RCMIP paper is the more comprehensive list -- e.g., "For a more extensive list of simple climate models, see Nicholls et al... (link)".
