### 2020-11-02 f6dfe3f
Clarify location of more detailed table
