### 2020-11-02 6233ce2
Fill out table with more openly accessible examples
