+++ paper.bib
@@ -173,3 +173,53 @@ DOI = {10.5194/gmd-11-2273-2018}
   URL = {https://gmd.copernicus.org/articles/8/939/2015/},
   DOI = {10.5194/gmd-8-939-2015}
 }
+
+@article{Meinshausen_2011_magicc,
+ author = {Meinshausen, M. and Raper, S. C. B. and Wigley, T. M. L.},
+ title = {Emulating coupled atmosphere-ocean and carbon cycle models with a simpler model, {MAGICC6} – Part 1: Model description and calibration},
+ journal = {Atmospheric Chemistry and Physics},
+ volume = {11},
+ year = {2011},
+ number = {4},
+ pages = {1417--1456},
+ doi = {10.5194/acp-11-1417-2011},
+}
+
+@article{Gieseke_2018_pymagicc,
+  doi = {10.21105/joss.00516},
+  url = {https://doi.org/10.21105/joss.00516},
+  year = {2018},
+  publisher = {The Open Journal},
+  volume = {3},
+  number = {22},
+  pages = {516},
+  author = {Robert Gieseke and Sven N. Willner and Matthias Mengel},
+  title = {Pymagicc: A Python wrapper for the simple climate model MAGICC},
+  journal = {Journal of Open Source Software}
+}
+
+@Article{Gasser_2020_asdfjk,
+  AUTHOR = {Gasser, T. and Crepin, L. and Quilcaille, Y. and Houghton, R. A. and Ciais, P. and Obersteiner, M.},
+  TITLE = {Historical \chem{CO_{2}} emissions from land use and land cover change and their
+  uncertainty},
+  JOURNAL = {Biogeosciences},
+  VOLUME = {17},
+  YEAR = {2020},
+  NUMBER = {15},
+  PAGES = {4075--4101},
+  URL = {https://bg.copernicus.org/articles/17/4075/2020/},
+  DOI = {10.5194/bg-17-4075-2020}
+}
+
+@article{Goodwin_2019_ggfp6s,
+    author = "Goodwin, Philip and Williams, Richard G. and Roussenov, Vassil M. and Katavouta, Anna",
+    doi = "10.1029/2019gl082887",
+    year = "2019",
+    month = "jul",
+    publisher = "American Geophysical Union ({AGU})",
+    volume = "46",
+    number = "13",
+    pages = "7554--7564",
+    title = "Climate Sensitivity From Both Physical and Carbon Cycle Feedbacks",
+    journal = "Geophysical Research Letters"
+}
