### 2020-11-02 #2766 @znicholls
> Oscar, WASP and Pymagicc/MAGICC

I've added these now

> Maybe one could reference models that are easily driven in forcing mode?

I could do this, that would mean I'd remove GREB. Seems fine to me as it's also kind of outside the typical idea of a simple climate model.
