### 2020-11-02 ca946f2
Update README and clarify text
