### 2020-11-02 #2766 @rgieseke
> >    Maybe one could reference models that are easily driven in forcing mode?
>
> I could do this, that would mean I'd remove GREB. Seems fine to me as it's also kind of outside the typical idea of a simple climate model.

There is probably no clear distinction, so i'd keep it. As it's also meant for teaching (i think) it would be a good pointer!
