### 2020-11-05 #2766 @sadielbartholomew
Hi, I can start my review in earnest shortly, so can I just check that the state of the repository and of the paper has stabilised now in terms of changes made resulting from feedback from the first review, @znicholls ?

It's been a few days since any activity (at least as indicated on this thread) so it seems likely, but think it best to check to avoid starting the review and have some change go in that might influence the judgement I can make.
