### 2020-11-05 e4da75d
Fix MAGICC link
