### 2020-11-05 #2766 @znicholls
> so can I just check that the state of the repository and of the paper has stabilised now in terms of changes made resulting from feedback from the first review, @znicholls ?

Yep I've just merged everything into master and will get whedon to recompile now. Thanks for checking and reviewing!
