### 2020-11-30 #2766 @sadielbartholomew
Hi @leouieda, thanks for checking in. Sorry for the delay, which has arisen because I've been prioritising another JOSS review which began at an earlier date that this one. But I have largely completed that now, so I can focus on this review. I'll aim to complete it by the end of the week. Thanks to all for your patience.
