+++ .github/workflows/test-pypi-install.yml
@@ -1,8 +1,9 @@
 name: Test PyPI install
-on:
-  schedule:
-    # * is a special character in YAML so you have to quote this string
-    - cron:  '0 * * * 3'
+on: push
+# on:
+#   schedule:
+#     # * is a special character in YAML so you have to quote this string
+#     - cron:  '0 * * * 3'
 
 jobs:
   test-installation:
@@ -18,8 +19,11 @@ jobs:
       with:
         python-version: ${{ matrix.python-version }}
     - name: Install package
+      # install basic first, then install pre-releases afterwards so we don't
+      # pick up pre-releases of dependencies (e.g. pandas)
       run: |
         pip install --upgrade pip
+        pip install openscm-twolayermodel
         pip install openscm-twolayermodel --pre
     - name: Checkout repository
       uses: actions/checkout@v2
