### 2020-12-11 #2766 @sadielbartholomew
Hi @leouieda @znicholls, I started my review however before proceeding with the remainder I feel obliged to request a more experienced opinion regarding the 'substantial scholarly effort' criterion because, while it is clear more than sufficient academic effort has been conducted related to the context of this package, I am not convinced the package itself meets the minimum threshold of three months work so, [as per the JOSS review guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort), I need to flag this so an informed decision can be made by the editors in this respect.

I say this tentatively, because it is difficult to grasp the effort put into a package. I am aware the model is simple by design and the fact it has been used and cited in the RCMIP paper indicate it is important to the research area and indeed very citable. My concerns stem largely from the line counts of the functional Python code, excluding tests and configuration scripts which seems to be ~1500 lines:

```console
$ pwd
/home/sadie/openscm-twolayermodel/src/openscm_twolayermodel
$ find . -name '*.py' | xargs wc -l
  314 ./base.py
   12 ./__init__.py
   15 ./errors.py
   10 ./constants.py
  321 ./impulse_response_model.py
  520 ./_version.py
  355 ./two_layer_model.py
   37 ./utils.py
 1584 total
```

which is low enough to plausibly have taken less than three months in total considering even the related research time for the idea, design, implementation planning and incremental changes, etc. The commit history indicates it has been worked on at least on GitHub for about two-thirds of a year, but the total work time can't be inferred from that.

@leouieda please can you provide the editorial opinion on whether I can consider this criterion met. If so I will be very happy to proceed with the remainder of the review.Thanks.
