### 2020-12-12 #2766 @znicholls
For context the comments in the pre-review:

mine: https://github.com/openjournals/joss-reviews/issues/2728#issuecomment-705214797
@leouieda's response at the time: https://github.com/openjournals/joss-reviews/issues/2728#issuecomment-707132474
