### 2020-12-15 #2766 @leouieda
@sadielbartholomew thank you for bringing this up. As @znicholls pointed out, we had discussed this in the prereview since the line count is on the small side. However, I'm not a climate researcher which is why we decided to get the opinion of reviewers as well. As you mentioned, this is evidently a useful piece of software given that it's already been cited. From my side, I would be happy to treat this as in scope if you agree. The "three months of effort" is more of a rule of thumb than a hard requirement. But as I said, I'm not an expert in this field. Please also consider the amount of effort that went into testing and documentation.

So disregarding the "3-month of effort" concept, do you think this qualifies as a "substantial scholarly effort"?
