### 2020-12-15 #2766 @ashiklom
My 2 cents: I think this is comfortably within scope. We as an Earth science community need more user-friendly modeling tools like this, and more citable papers describing those tools.
