### 2020-12-15 #2766 @rgieseke
I'm obiously biased (as i'm part of the OpenSCM projects) but this tool is also using the `scmdata` and `openscm-units` projects by the same authors which help making the code in this project look succinct by moving boilerplate code to re-usable modules..
