### 2020-12-15 #2766 @sadielbartholomew
Thanks all for the comments RE the 'substantial scholarly effort' criterion.

> However, I'm not a climate researcher which is why we decided to get the opinion of reviewers as well.

I work as an effective research software engineer in the field of climate research, however that's quite different to being a climate researcher, and I am definitely not an "expert" in the field either! So that's an important caveat to note. If you require the opinion of someone who fits that kind of description, I do have numerous colleagues who would fit that description, some of whom I might be able to seek the opinion of (should that be permitted relating to COI rules).

However, if my own opinion is sufficient, based on the comments here, in particular that our first reviewer @ashiklom regards it as "comfortably within scope", that my answer to the following is 'yes, absolutely':

> So disregarding the "3-month of effort" concept, do you think this qualifies as a "substantial scholarly effort"?

and the good point by @rgieseke that:

> this tool is also using the scmdata and openscm-units projects by the same authors which help making the code in this project look succinct by moving boilerplate code to re-usable modules

where fewer lines of code is a mark of quality in my eyes, I am more than happy to consider OpenSCM Two Layer Model as being 'substantial scholarly effort'.

Are we okay to take my own judgement (given the above caveat) @leouieda? (If so, I will tick that item and I will finish my review soon.) If not, would it help for me to check in with a colleague that is an experienced climate researcher (who can self-certify as having no COI) for their view?

As a side note, I do generally agree with the sentiments such as:

> We as an Earth science community need more user-friendly modeling tools like this, and more citable papers describing those tools.

but I am trying to be as objective as possible and judge purely on the review criteria.
