### 2020-12-23 bd3b559
Update to requiring scmdata with new openscm-units
