### 2020-12-23 cc8dc17
Remove pip 2020 feature resolver
