### 2020-12-23 89108d6
Fix test notebook command [ci skip]
