### 2021-01-06 #2766 @leouieda
@sadielbartholomew thank you for being so thorough and trying to look at this from a different perspective. I'm happy to take your own judgement on this since @ashiklom agrees. Please let me know when you're done and all checks have been marked in the review issue.

@znicholls thank you for your patience since this review has taken longer than usual, in no small part because of me. After reviews are done, there are some editorial procedures to go through but they should be fairly quick.

Happy New Year to everyone! 🍾
