### 2021-03-08 #2766 @sadielbartholomew
## Summary of review

I am very satisfied with both the software and the paper, notably that they meet the requirements. Therefore I have now ticked all review list checkboxes and I am happy to recommend acceptance of the OpenSCM Two Layer Model paper for publication.

Please do consider the minor comments I raise below @znicholls, though they are merely suggestions or questions that came to me during review and whether or not any changes are made in response to them is up to you, it does not influence my decision to suggest approving the paper to @leouieda.

The most important of these comments are probably those regarding the documentation of, and output of, the the automated tests (I'm about to open some minor PRs regarding these items).

Thanks to all for awaiting my review.

*Note*: for impartiality I have not re-read through the comments of the first reviewer before my review, and hence not referenced them here. Therefore I may have covered points that have already been raised. If they have been addressed satisfactorily already, there is no need for further action.

## Minor comments

Organised in correspondence to the review checklist (no mention of a given check box item means I don't have any comments to make about it, other than that I am satisfied it is met by the software and/or paper):

### Functionality

* '**Functionality**': there are a few potential issues raised by some failures and warnings from a run of the tests: see the 'Documentation' -> 'Automated tests' section below for details.

* ['**Performance**': just to note that I found no performance claims, rather than having found some and confirmed they are met.]

### Documentation

* '**Installation instructions**': installation is handled seamlessly with both `pip` and `conda`, and the corresponding package dependencies are [enumerated in the `setup.py`](https://github.com/openscm/openscm-twolayermodel/blob/19b030571892a3238082765671e161ddd4c2ab97/setup.py#L21-L68), though I think for potential users it would be useful to additionally add, or process in, the list of the user-facing (i.e. install rather than development) dependencies to the 'Installation' section of the docs, in case pontential users who are not sure if they want to try out the software want to know the requirements, without having to search for them (or some similar scenario).

* '**Automated tests**': though there clearly are tests living under the `tests` repo, I struggled to find information about the standard ways to run them either in the docs or in the README. After searching and  effort than is ideal, I eventually found "Make sure the tests pass by running make check" under ['Getting setup'](https://openscm-two-layer-model.readthedocs.io/en/stable/development.html?highlight=tests#getting-setup) but that didn't work (gave `make: *** No rule to make target 'check'. Stop.`) and only by inspecting the `Makefile` did I see that the command should in fact be `make checks`, so finally I managed to run the tests.

    I suggest adding an explicit instruction about running the tests via this command when in the root of the directory (or otherwise) to the 'installation' page. It would be helpful to add it to the repo README as well under a section called 'Tests' or similar.

    Moreover when I ran the checks on the current `master` branch there were 4 failures and 2 signficant warnings, so it looks like there are some self-contained issue to be resolved.

    I am about to open a few PRs regarding these (I will link them to this Issue).

* '**Community guidelines**': the guidelines are clear, with (1) covered particularly thoroughly, though I wonder whether guidance for (2) and (3) might be somewhat difficult to find in the case of a user looking for support, because the guidelines appear to sit under the 'Development' section of the documentation (unless I have missed some further guidance given elsewhere in the docs?). Ideally, an indication of how to get help could be provided outside of the parts of the docs which might be considered to apply to developers only.

### Software paper

* '**Summary**': very good, clear description. However, I did note that the software was described as ' object-oriented' and thought that the paper could be enhanced with a little description (or perhaps a schematic) of the design of OpenSCM Two Layer Model in terms of objects and their relationships, e.g. the use of the model as an abstract base class, a two-layer variant and model sub-classing from it, etc.

* '**State of the field**': several other commonly-used packages are *mentioned* and *described*, though it appears that only the FaIR model is *explicitly compared* to the OpenSCM Two Layer Model in the text. Ideally the other models in the table could be briefly compared too, for the benefit of non-specialist readers who would not know how the brief description of the approaches, also in the table, differentiate between the simple open models.

* '**Quality of writing**': very well written, and there is nothing I can see that should be edited in any way, except a few formatting aspects that in my opinion could make the table clearer:
    1. The reporting of the repositories in the 'Programming language' column clutters the table, making it difficult to pick out the language in question at a glance, in my opinion. Plus it seems the relevant URLs are already hyperlinked to the model names in the model column. I suggest either:
        * moving the parenthesised codebase URLs into the 'Model' column and hyperlinking them there;
         * hyperlinking the codebase URLs instead of the model names so the different face for the links makes the language item stand out in that column.
    2. In the per-repository language breakdown GitHub is suggesting some of the models in the table are not fully written in the language quoted under 'Programming language', e.g. it indicates significant MATLAB code along with the predominnant C++ in WASP (though it is not clear if the MATLAB is part of the model or some extra resource in the repo). Also some of the models are stated to have wrappers available for other languages than the one quoted. So perhaps the title could be pre-fixed with 'main', 'original', 'reference' or similar ('Main programming language' etc.) to indicate this?

* '**References**': the paper is well-referenced and the format of the references is sound, but I did notice there are two sets of papers cited which have the same authors and publication date, namely "Geoffroy, Saint-Martin, Bellon, et al., 2013" and "M. Meinshausen et al., 2011". Ideally these could be distinguished in the references in the text, somehow, else someone reading would not be able to tell which of the two was being referenced, especially if reading a printed copy. Perhaps "M. Meinshausen et al., 2011a" and "...200b" as suggested here?

    There doesn't seem to be a way I can find quickly on how to do this with the JOSS standard citation syntax, CSL. @leouieda do you know how it could be done?
