### 2021-03-08 2bf866d
Fix commands under 'getting setup' in development documentation
