### 2021-04-06 #2766 @danielskatz
👋 @openjournals/dev - see the last comment (References) in https://github.com/openjournals/joss-reviews/issues/2766#issuecomment-792447820 - is there a good answer to this?
