### 2021-04-23 #2766 @arfon
> 👋 @openjournals/dev - see the last comment (References) in #2766 (comment) - is there a good answer to this?

I'm afraid I don't know sorry. Perhaps @tarleb could provide some pointers?
