### 2021-04-26 #2766 @znicholls
Progress towards addressing the comments from @sadielbartholomew's excellent review (https://github.com/openjournals/joss-reviews/issues/2766#issuecomment-792447820):

- [ ] warnings in tests
- [ ] add dependencies to installation section of the docs
- [ ] fix docs about running tests (makefile command is wrong)
- [ ] add section in docs about running tests to installation page and in README
- [ ] fix tests
- [ ] clearer section on where to get help
- [ ] re-do the table
- [ ] briefly compare other models to what is implemented here
