+++ paper.md
@@ -43,14 +43,24 @@ On top of the relationship between ocean heat uptake and warming, these models a
 The exception is the FaIR model [@smith_2018_fairv1_3], which uses the two-layer model as its thermal core.
 For a more extensive list of simple climate models and references which describe the models in detail, see Table 1 of @rcmip_phase_1.
 
-| Model | Brief description | Programming language |
-|-------|-------------------|----------------------|
-| [FaIR](https://github.com/OMS-NetZero/FAIR) | Modified impulse response [@smith_2018_fairv1_3] | Python (github.com/OMS-NetZero/FAIR) |
-| [GREB](https://github.com/christianstassen/greb-official) | Coarse grid energy balance [@Dommenget_2011_greb] | Fortran 90 (github.com/christianstassen/greb-official) |
-| [Hector](https://github.com/JGCRI/hector) | Upwelling-diffusion ocean energy balance [@hartin_2015_hector] | C++ (github.com/JGCRI/hector) |
-| [MAGICC](http://magicc.org) | Upwelling-diffusion ocean four-box (hemispheric land/ocean) energy balance [@Meinshausen_2011_magicc] | Fortran 90 (live.magicc.org, Pymagicc [@Gieseke_2018_pymagicc] provides a Python wrapper at github.com/openclimatedata/pymagicc) |
-| [OSCAR](https://github.com/tgasser/OSCAR) | Energy balance with book-keeping land carbon cycle [@Gasser_2020_asdfjk] | Python (github.com/tgasser/OSCAR) |
-| [WASP](https://github.com/WASP-ESM/WASP_Earth_System_Model) | Energy balance with 8-box carbon cycle [@Goodwin_2019_ggfp6s] | C++ (github.com/WASP-ESM/WASP_Earth_System_Model) |
++-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
+| Model                                                       | Brief description and URL                                                                                                                                                                                                   | Language                    |
+|                                                             |                                                                                                                                                                                                                             |                              |
++=============================================================+=============================================================================================================================================================================================================================+==============================+
+| [FaIR](https://github.com/OMS-NetZero/FAIR)                 | Modified impulse response [@smith_2018_fairv1_3], github.com/OMS-NetZero/FAIR                                                                                                                                               | Python                       |
++-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
+| [GREB](https://github.com/christianstassen/greb-official)   | Coarse grid energy balance [@Dommenget_2011_greb], github.com/christianstassen/greb-official                                                                                                                                | Fortran 90                   |
++-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
+| [Hector](https://github.com/JGCRI/hector)                   | Upwelling-diffusion ocean energy balance [@hartin_2015_hector], github.com/JGCRI/hector                                                                                                                                     | C++                          |
++-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
+| [MAGICC](http://magicc.org)                                 | Upwelling-diffusion ocean four-box (hemispheric land/ocean) energy balance [@Meinshausen_2011_magicc], live.magicc.org (Pymagicc [@Gieseke_2018_pymagicc] provides a Python wrapper at github.com/openclimatedata/pymagicc) | Fortran 90                   |
[…]
