### 2021-04-26 c268324
Fix keyword duplication in tests
