### 2021-04-26 345489f
Add section in docs about running tests
