+++ docs/source/development.rst
@@ -1,10 +1,10 @@
-.. development:
+.. _development:
 
 Development
 ===========
 
 If you're interested in contributing to OpenSCM Two Layer Model, we'd love to have you on board!
-This section of the docs will (once we've written it) detail how to get setup to contribute and how best to communicate.
+This section of the docs details how to get setup to contribute and how best to communicate.
 
 .. contents:: :local:
 
