+++ docs/source/index.rst
@@ -14,6 +14,14 @@ OpenSCM Two Layer Model
     :start-after: sec-begin-license
     :end-before: sec-end-license
 
+.. include:: ../../README.rst
+    :start-after: sec-begin-getting-help
+    :end-before: sec-end-getting-help
+
+.. include:: ../../README.rst
+    :start-after: sec-begin-links
+    :end-before: sec-end-links
+
 .. toctree::
     :maxdepth: 2
     :caption: Documentation
