+++ docs/source/index.rst
@@ -18,10 +18,6 @@ OpenSCM Two Layer Model
     :start-after: sec-begin-getting-help
     :end-before: sec-end-getting-help
 
-.. include:: ../../README.rst
-    :start-after: sec-begin-links
-    :end-before: sec-end-links
-
 .. toctree::
     :maxdepth: 2
     :caption: Documentation
