+++ .github/workflows/ci-cd-workflow.yml
@@ -48,8 +48,6 @@ jobs:
           fi
 
   build:
-    needs: linting-and-docs
-
     runs-on: ubuntu-latest
     strategy:
       matrix:
@@ -79,7 +77,6 @@ jobs:
           fi
 
   test-notebooks:
-    needs: build
     runs-on: ubuntu-latest
     strategy:
       matrix:
@@ -104,8 +101,25 @@ jobs:
       run: |
         pytest docs/source/usage -r a --nbval --sanitize-with tests/notebook-tests.cfg --no-cov
 
+  readme-requirements-up-to-date:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [3.7]
+
+    steps:
+    - name: Checkout repository
+      uses: actions/checkout@v2
+    - name: Setup python
+      uses: actions/setup-python@v1
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Check README is up-to-date with setup.py
+      run: |
+          python scripts/check_requirements_in_readme.py
+
   deploy-pypi:
-    needs: test-notebooks
+    needs: build,linting-and-docs,readme-requirements-up-to-date,test-notebooks
     if: startsWith(github.ref, 'refs/tags/v')
 
     runs-on: ubuntu-latest
