+++ .github/workflows/ci-cd-workflow.yml
@@ -119,7 +119,7 @@ jobs:
           python scripts/check_requirements_in_readme.py
 
   deploy-pypi:
-    needs: build,linting-and-docs,readme-requirements-up-to-date,test-notebooks
+    needs: [build,linting-and-docs,readme-requirements-up-to-date,test-notebooks]
     if: startsWith(github.ref, 'refs/tags/v')
 
     runs-on: ubuntu-latest
