### 2021-04-26 91e79cc
Remove unknown command
