+++ paper.bib
@@ -200,7 +200,7 @@ DOI = {10.5194/gmd-11-2273-2018}
 
 @Article{Gasser_2020_asdfjk,
   AUTHOR = {Gasser, T. and Crepin, L. and Quilcaille, Y. and Houghton, R. A. and Ciais, P. and Obersteiner, M.},
-  TITLE = {Historical \chem{CO_{2}} emissions from land use and land cover change and their
+  TITLE = {Historical {CO$_2$} emissions from land use and land cover change and their
   uncertainty},
   JOURNAL = {Biogeosciences},
   VOLUME = {17},
