### 2021-04-26 #2766 @tarleb
> I'm afraid I don't know sorry. Perhaps @tarleb could provide some pointers?

Looks like a pandoc bug, possibly jgm/citeproc#44. I haven't tried yet, but maybe updating pandoc would fix this?
