### 2021-05-03 #2766 @znicholls
Alright I'm all done replying to the second review (https://github.com/openjournals/joss-reviews/issues/2766#issuecomment-792447820), see checklist here https://github.com/openjournals/joss-reviews/issues/2766#issuecomment-826437944 :)
