### 2021-05-04 ce00f59
Add Python 3.9 tests
