+++ .github/workflows/ci-cd-workflow.yml
@@ -66,9 +66,9 @@ jobs:
         pip install -e .[tests]
     - name: Test with pytest
       env:
-        MIN_COVERAGE: 70  # TODO: put this back up to 95 once we start writing
+        MIN_COVERAGE: 95
       run: |
-          pytest tests -r a --cov=openscm_twolayermodel --cov-report=''
+          pytest tests -r a --cov=openscm_twolayermodel --cov-report=xml
           if ! coverage report --fail-under=${MIN_COVERAGE} --show-missing
           then
               echo
