### 2021-05-04 0126ece
Add conda install test
