### 2021-05-04 3f2e720
Put conda and pypi install tests back on schedule
