+++ README.rst
@@ -10,7 +10,7 @@ OpenSCM Two Layer Model
 +---------------+--------+--------+
 
 +------+------------------+----------------+------------------+
-| Pypi |  |PyPI Install|  |     |PyPI|     |  |PyPI Version|  |
+| PyPI |  |PyPI Install|  |     |PyPI|     |  |PyPI Version|  |
 +------+------------------+----------------+------------------+
 
 +-------+-----------------+-------------------+-----------------+
