### 2021-05-04 eb10bf2
Remove Python 3.6 from conda tests
