+++ .github/workflows/test-conda-install.yml
@@ -1,8 +1,5 @@
 name: Test conda install
-on:
-  schedule:
-    # * is a special character in YAML so you have to quote this string
-    - cron:  '0 * * * 3'
+on: push
 
 jobs:
   test-installation:
