### 2021-05-06 #2766 @sadielbartholomew
Thanks for addressing my comments so thoroughly @znicholls. For confirmation that nothing has inadvertently been changed for the worse when making those improvements I have re-read through the paper and made subsequent basic checks on the codebase (re-running the tests locally, eyeballing some core pages of the current documentation etc.) and everything is great with no issues to report at this stage, except perhaps as a final minor note (I am not sure if you noticed this yourself) that I saw a sentence is now repeated twice in the paper in the main text (lines 33-35) and at the end of the caption for Table 1:

> For a more extensive list of simple climate models and references which describe the models in detail, see Table 1 of Nicholls et al. (2020).

and for conciseness you may prefer to remove one instance.

@leouieda as far as I can see both @ashiklom and myself are very happy with this submission and it is ready to go so I believe we can hand it over to you for next steps, etc.?
