### 2021-06-11 #2766 @leouieda
@znicholls I see that the latest version of the paper is in the `paper-typos` branch. I'll my proof-reading on that version. Afterwards, it would be good to merge everything into `master`.
