### 2021-06-11 a56424b
Use markdown links in the table of models
Placing the URLs doesn't render them as links in the paper.
Use `[]()` to make them clickable.
