+++ paper.md
@@ -48,17 +48,17 @@ The exception is the FaIR model [@smith_2018_fairv1_3], which uses the two-layer
 | Model                                                       | Brief description and URL                                                                                                                                                                                                   | Language                    |
 |                                                             |                                                                                                                                                                                                                             |                              |
 +=============================================================+=============================================================================================================================================================================================================================+==============================+
-| [FaIR](https://github.com/OMS-NetZero/FAIR)                 | Modified impulse response [@smith_2018_fairv1_3], github.com/OMS-NetZero/FAIR                                                                                                                                               | Python                       |
+| [FaIR](https://github.com/OMS-NetZero/FAIR)                 | Modified impulse response [@smith_2018_fairv1_3], [github.com/OMS-NetZero/FAIR](https://github.com/OMS-NetZero/FAIR)                                                                                                                                               | Python                       |
 +-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
-| [GREB](https://github.com/christianstassen/greb-official)   | Coarse grid energy balance [@Dommenget_2011_greb], github.com/christianstassen/greb-official                                                                                                                                | Fortran 90                   |
+| [GREB](https://github.com/christianstassen/greb-official)   | Coarse grid energy balance [@Dommenget_2011_greb], [github.com/christianstassen/greb-official](https://github.com/christianstassen/greb-official)                                                                                                                                | Fortran 90                   |
 +-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
-| [Hector](https://github.com/JGCRI/hector)                   | Upwelling-diffusion ocean energy balance [@hartin_2015_hector], github.com/JGCRI/hector                                                                                                                                     | C++                          |
+| [Hector](https://github.com/JGCRI/hector)                   | Upwelling-diffusion ocean energy balance [@hartin_2015_hector], [github.com/JGCRI/hector](https://github.com/JGCRI/hector)                                                                                                                                     | C++                          |
 +-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
-| [MAGICC](http://magicc.org)                                 | Upwelling-diffusion ocean four-box (hemispheric land/ocean) energy balance [@Meinshausen_2011_magicc], live.magicc.org (Pymagicc [@Gieseke_2018_pymagicc] provides a Python wrapper at github.com/openclimatedata/pymagicc) | Fortran 90                   |
+| [MAGICC](http://magicc.org)                                 | Upwelling-diffusion ocean four-box (hemispheric land/ocean) energy balance [@Meinshausen_2011_magicc], [http://live.magicc.org](live.magicc.org) (Pymagicc [@Gieseke_2018_pymagicc] provides a Python wrapper at [github.com/openclimatedata/pymagicc](https://github.com/openclimatedata/pymagicc)) | Fortran 90                   |
 +-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
-| [OSCAR](https://github.com/tgasser/OSCAR)                   | Energy balance with book-keeping land carbon cycle [@Gasser_2020_asdfjk], github.com/tgasser/OSCAR                                                                                                                          | Python                       |
[…]
