+++ paper.md
@@ -54,7 +54,7 @@ The exception is the FaIR model [@smith_2018_fairv1_3], which uses the two-layer
 +-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
 | [Hector](https://github.com/JGCRI/hector)                   | Upwelling-diffusion ocean energy balance [@hartin_2015_hector], [github.com/JGCRI/hector](https://github.com/JGCRI/hector)                                                                                                                                     | C++                          |
 +-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
-| [MAGICC](http://magicc.org)                                 | Upwelling-diffusion ocean four-box (hemispheric land/ocean) energy balance [@Meinshausen_2011_magicc], [http://live.magicc.org](live.magicc.org) (Pymagicc [@Gieseke_2018_pymagicc] provides a Python wrapper at [github.com/openclimatedata/pymagicc](https://github.com/openclimatedata/pymagicc)) | Fortran 90                   |
+| [MAGICC](http://magicc.org)                                 | Upwelling-diffusion ocean four-box (hemispheric land/ocean) energy balance [@Meinshausen_2011_magicc], [live.magicc.org](http://live.magicc.org) (Pymagicc [@Gieseke_2018_pymagicc] provides a Python wrapper at [github.com/openclimatedata/pymagicc](https://github.com/openclimatedata/pymagicc)) | Fortran 90                   |
 +-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
 | [OSCAR](https://github.com/tgasser/OSCAR)                   | Energy balance with book-keeping land carbon cycle [@Gasser_2020_asdfjk], [github.com/tgasser/OSCAR](https://github.com/tgasser/OSCAR)                                                                                                                          | Python                       |
 +-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
