+++ paper.md
@@ -44,30 +44,22 @@ These models are conceptually similar to the two-layer model implemented here ex
 On top of the relationship between ocean heat uptake and warming, these models also implement many other components of the climate system, e.g. carbon cycle, methane cycle and the relationship between changes in atmospheric greenhouse gas concentrations and atmospheric energy fluxes.
 The exception is the FaIR model [@smith_2018_fairv1_3], which uses the two-layer model as its thermal core.
 
-+-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
-| Model                                                       | Brief description and URL                                                                                                                                                                                                   | Language                    |
-|                                                             |                                                                                                                                                                                                                             |                              |
-+=============================================================+=============================================================================================================================================================================================================================+==============================+
-| [FaIR](https://github.com/OMS-NetZero/FAIR)                 | Modified impulse response [@smith_2018_fairv1_3], [github.com/OMS-NetZero/FAIR](https://github.com/OMS-NetZero/FAIR)                                                                                                                                               | Python                       |
-+-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
-| [GREB](https://github.com/christianstassen/greb-official)   | Coarse grid energy balance [@Dommenget_2011_greb], [github.com/christianstassen/greb-official](https://github.com/christianstassen/greb-official)                                                                                                                                | Fortran 90                   |
-+-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
-| [Hector](https://github.com/JGCRI/hector)                   | Upwelling-diffusion ocean energy balance [@hartin_2015_hector], [github.com/JGCRI/hector](https://github.com/JGCRI/hector)                                                                                                                                     | C++                          |
-+-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
-| [MAGICC](http://magicc.org)                                 | Upwelling-diffusion ocean four-box (hemispheric land/ocean) energy balance [@Meinshausen_2011_magicc], [live.magicc.org](http://live.magicc.org) (Pymagicc [@Gieseke_2018_pymagicc] provides a Python wrapper at [github.com/openclimatedata/pymagicc](https://github.com/openclimatedata/pymagicc)) | Fortran 90                   |
-+-------------------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+
-| [OSCAR](https://github.com/tgasser/OSCAR)                   | Energy balance with book-keeping land carbon cycle [@Gasser_2020_asdfjk], [github.com/tgasser/OSCAR](https://github.com/tgasser/OSCAR)                                                                                                                          | Python                       |
[…]
