+++ paper.md
@@ -12,6 +12,7 @@ authors:
     orcid: 0000-0003-0872-7098
     affiliation: "1, 2, 3"
   - name: Jared Lewis
+    orcid: 0000-0002-8155-8924
     affiliation: "1, 2, 3"
 affiliations:
  - name: Australian-German Climate & Energy College, The University of Melbourne, Parkville, Victoria, Australia
