### 2021-06-14 #2766 @leouieda
@znicholls alright, I'm very happy with the state of the paper and the reviewers are satisfied as well 🎉

There just a few of things that you'll need to dd:

1. Double check authors and affiliations (including ORCIDs)
2. Make a release of the software with the latest changes from the review. This is the version that will be used in the JOSS paper.  **Post the version number here** once it's available.
3. Archive the release on Zenodo/figshare/etc and **post the DOI here**.
4. Make sure that the **title and author list** (including ORCIDs) in the archive match those in the JOSS paper.

Once those are done, we can move on to publication! Let me know if you need any help with these or have any questions.
