### 2021-06-15 #2766 @znicholls
All done :)

- [x] Double check authors and affiliations (including ORCIDs) (https://github.com/openscm/openscm-twolayermodel/pull/36)
- [x] Make a release of the software with the latest changes from the review. 'v0.2.3'
- [x] Archive the release on Zenodo/figshare/etc and post the DOI here. https://doi.org/10.5281/zenodo.4950772
