### 2021-06-15 #2766 @leouieda
@znicholls I'm happy to recommend acceptance of this paper for publication in JOSS!  🎉 🎊

Thank you to our reviewers @sadielbartholomew and @ashiklom for your thorough and helpful comments! I'm particularly grateful for all of your patience and understanding with the big delay in this review.
