### 2021-06-15 #2766 @danielskatz
Thanks @znicholls - this basically looks good to me to finalize, though I've suggested some small changes in the paper in https://github.com/openscm/openscm-twolayermodel/pull/37  Please merge this, or let me know what you disagree with, then we can proceed to publishing.
